Paginate search and list items responses (#272)
* Remove existing pagination strategy

* Downgrade jdk to 8 for memory reasons maybe :thinking:

In other projects, metals memory usage explodes in jdk 11. In GeoTrellis and
RF we build on JDK 8. It's possible that our CI memory issues are related to
the JDK 11 container and nothing in particular to worry about other than that.

* Add serial_id and created_at columns to collection_items

* Add PaginationToken and test its schema

* Add pagintation parameters to both list endpoints

* fix api test compilation

* Paginate item list endpoint

* Paginate search responses

* lint all the things

* Handle pagination token filtering consistently

* Add serial and timestamp to collections

This matters because I needed to modify the listQ in Dao, which
everything extends, which means that non-item endpoints were going to
break without this change. Collections aren't currently paginated, but
they could be someday, so we'll have this lying around for if we
need it and the payoff now is that we don't need special logic per DAO

* Include next request information in POST searches

* remove accidental banner change

* include link to paging extension source

* ci help -- don't run tests in parallel

* use query parameter conversion from SearchFilters

* constrain container sizes

* lint

* use existing and good bboxToString method

* lint

* restore parallel tests

WILL IT SURVIVE

i hope so

i'll be sad if it doesn't

* ignore pagination when counting matching results

* only return next link if there are more results