Decode requested extensions on serialization (#285)
* wip

* do validation on item list serialization

* don't blow up if no extensions passed

* add tests that fail using stac4s testing

* lint

* slightly more lint

* remove debugging printlns

* make decodingerror serde more cooperative

This isn't a perfect solution, but Circe doesn't want to report
the issue that it runs into anyway. There's an open issue from four
years ago in Circe to pretty print the history better that we could
consider, but for now we're just reporting every field that the decoder
thinks has gone poorly. This also leads to some unexpected results with
`label:`, where the decoder in stac4s wants more specific things than
I think is technically valid (or maybe our example catalog in S3 is
just not quite right :thinking:). Here's the circe issue:

https://github.com/circe/circe/issues/188

* use stac4s testing module for stac4s arbitraries

* always validate items on collections/<id>/items

* lint

* fix tests